Move Factory into helpers folder
