+++ src/MetroidInvaders.cpp
@@ -1,8 +1,8 @@
 #include "../include/MetroidInvaders.hpp"
-#include "../include/Factory.hpp"
 #include "../include/Title.hpp"
 #include "../include/Game.hpp"
 #include "../include/GameOver.hpp"
+#include "../include/helpers/Factory.hpp"
 #include "../include/adaptors/RenderWindowAdaptor.hpp"
 #include "../include/adaptors/ClockAdaptor.hpp"
 #include "../include/interfaces/collisionInterface.hpp"
